EncryptoBot/requirements.txt
aiogram~=2.25
pynacl~=1.5.0
pycparser~=2.21
